apps/api/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-generativeai
requests
pytrends
ddgs
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy